third_party/WebKit/LayoutTests/imported/wpt/html/dom/reflection-forms-expected.txt

CONSOLE WARNING: line 559: The <keygen> element is deprecated and will be removed in M56, around January 2017. See https://www.chromestatus.com/features/5716060992962560 for more details.
This is a testharness.js-based test.
PASS form.title: 77 tests
PASS form.lang: 77 tests
PASS form.dir: 151 tests
PASS form.className (<form class>): 77 tests
PASS form.hidden: 79 tests
PASS form.accessKey: 77 tests
PASS form.tabIndex: 10 tests
FAIL form.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS form.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL form.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS form.tabIndex: 36 tests
FAIL form.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS form.tabIndex: 2 tests
FAIL form.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS form.acceptCharset (<form accept-charset>): 77 tests
PASS form.autocomplete: 126 tests
PASS form.enctype: 151 tests
PASS form.encoding (<form enctype>): 151 tests
PASS form.method: 151 tests
PASS form.name: 77 tests
PASS form.noValidate: 79 tests
PASS form.target: 77 tests
PASS fieldset.title: 77 tests
PASS fieldset.lang: 77 tests
PASS fieldset.dir: 151 tests
PASS fieldset.className (<fieldset class>): 77 tests
PASS fieldset.hidden: 79 tests
PASS fieldset.accessKey: 77 tests
PASS fieldset.tabIndex: 10 tests
FAIL fieldset.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS fieldset.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL fieldset.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS fieldset.tabIndex: 36 tests
FAIL fieldset.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS fieldset.tabIndex: 2 tests
FAIL fieldset.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS fieldset.disabled: 79 tests
PASS fieldset.name: 77 tests
PASS legend.title: 77 tests
PASS legend.lang: 77 tests
PASS legend.dir: 151 tests
PASS legend.className (<legend class>): 77 tests
PASS legend.hidden: 79 tests
PASS legend.accessKey: 77 tests
PASS legend.tabIndex: 10 tests
FAIL legend.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS legend.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL legend.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS legend.tabIndex: 36 tests
FAIL legend.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS legend.tabIndex: 2 tests
FAIL legend.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS legend.align: 77 tests
PASS label.title: 77 tests
PASS label.lang: 77 tests
PASS label.dir: 151 tests
PASS label.className (<label class>): 77 tests
PASS label.hidden: 79 tests
PASS label.accessKey: 77 tests
PASS label.tabIndex: 10 tests
FAIL label.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS label.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL label.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS label.tabIndex: 36 tests
FAIL label.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS label.tabIndex: 2 tests
FAIL label.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS label.htmlFor (<label for>): 77 tests
PASS input.title: 77 tests
PASS input.lang: 77 tests
PASS input.dir: 151 tests
PASS input.className (<input class>): 77 tests
PASS input.hidden: 79 tests
PASS input.accessKey: 77 tests
PASS input.tabIndex: 10 tests
FAIL input.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS input.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL input.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS input.tabIndex: 36 tests
FAIL input.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS input.tabIndex: 2 tests
FAIL input.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS input.accept: 77 tests
PASS input.alt: 77 tests
PASS input.autofocus: 79 tests
PASS input.defaultChecked (<input checked>): 79 tests
PASS input.dirName: 77 tests
PASS input.disabled: 79 tests
PASS input.formEnctype: 151 tests
PASS input.formMethod: 126 tests
PASS input.formNoValidate: 79 tests
PASS input.formTarget: 77 tests
PASS input.inputMode: 5 tests
FAIL input.inputMode: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " followed by IDL get assert_equals: expected "" but got " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo "
PASS input.inputMode: setAttribute() to undefined followed by getAttribute() 
FAIL input.inputMode: setAttribute() to undefined followed by IDL get assert_equals: expected "" but got "undefined"
PASS input.inputMode: setAttribute() to 7 followed by getAttribute() 
FAIL input.inputMode: setAttribute() to 7 followed by IDL get assert_equals: expected "" but got "7"
PASS input.inputMode: setAttribute() to 1.5 followed by getAttribute() 
FAIL input.inputMode: setAttribute() to 1.5 followed by IDL get assert_equals: expected "" but got "1.5"
PASS input.inputMode: setAttribute() to true followed by getAttribute() 
FAIL input.inputMode: setAttribute() to true followed by IDL get assert_equals: expected "" but got "true"
PASS input.inputMode: setAttribute() to false followed by getAttribute() 
FAIL input.inputMode: setAttribute() to false followed by IDL get assert_equals: expected "" but got "false"
PASS input.inputMode: setAttribute() to object "[object Object]" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to object "[object Object]" followed by IDL get assert_equals: expected "" but got "[object Object]"
PASS input.inputMode: setAttribute() to NaN followed by getAttribute() 
FAIL input.inputMode: setAttribute() to NaN followed by IDL get assert_equals: expected "" but got "NaN"
PASS input.inputMode: setAttribute() to Infinity followed by getAttribute() 
FAIL input.inputMode: setAttribute() to Infinity followed by IDL get assert_equals: expected "" but got "Infinity"
PASS input.inputMode: setAttribute() to -Infinity followed by getAttribute() 
FAIL input.inputMode: setAttribute() to -Infinity followed by IDL get assert_equals: expected "" but got "-Infinity"
PASS input.inputMode: setAttribute() to "\0" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "\0" followed by IDL get assert_equals: expected "" but got "\0"
PASS input.inputMode: setAttribute() to null followed by getAttribute() 
FAIL input.inputMode: setAttribute() to null followed by IDL get assert_equals: expected "" but got "null"
PASS input.inputMode: setAttribute() to object "test-toString" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to object "test-toString" followed by IDL get assert_equals: expected "" but got "test-toString"
PASS input.inputMode: setAttribute() to object "test-valueOf" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to object "test-valueOf" followed by IDL get assert_equals: expected "" but got "test-valueOf"
PASS input.inputMode: 3 tests
FAIL input.inputMode: setAttribute() to "xverbatim" followed by IDL get assert_equals: expected "" but got "xverbatim"
PASS input.inputMode: setAttribute() to "verbatim\0" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "verbatim\0" followed by IDL get assert_equals: expected "" but got "verbatim\0"
PASS input.inputMode: setAttribute() to "erbatim" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "erbatim" followed by IDL get assert_equals: expected "" but got "erbatim"
PASS input.inputMode: setAttribute() to "VERBATIM" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "VERBATIM" followed by IDL get assert_equals: expected "verbatim" but got "VERBATIM"
PASS input.inputMode: 3 tests
FAIL input.inputMode: setAttribute() to "xlatin" followed by IDL get assert_equals: expected "" but got "xlatin"
PASS input.inputMode: setAttribute() to "latin\0" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "latin\0" followed by IDL get assert_equals: expected "" but got "latin\0"
PASS input.inputMode: setAttribute() to "atin" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "atin" followed by IDL get assert_equals: expected "" but got "atin"
PASS input.inputMode: setAttribute() to "LATIN" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "LATIN" followed by IDL get assert_equals: expected "latin" but got "LATIN"
PASS input.inputMode: 3 tests
FAIL input.inputMode: setAttribute() to "xlatin-name" followed by IDL get assert_equals: expected "" but got "xlatin-name"
PASS input.inputMode: setAttribute() to "latin-name\0" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "latin-name\0" followed by IDL get assert_equals: expected "" but got "latin-name\0"
PASS input.inputMode: setAttribute() to "atin-name" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "atin-name" followed by IDL get assert_equals: expected "" but got "atin-name"
PASS input.inputMode: setAttribute() to "LATIN-NAME" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "LATIN-NAME" followed by IDL get assert_equals: expected "latin-name" but got "LATIN-NAME"
PASS input.inputMode: 3 tests
FAIL input.inputMode: setAttribute() to "xlatin-prose" followed by IDL get assert_equals: expected "" but got "xlatin-prose"
PASS input.inputMode: setAttribute() to "latin-prose\0" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "latin-prose\0" followed by IDL get assert_equals: expected "" but got "latin-prose\0"
PASS input.inputMode: setAttribute() to "atin-prose" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "atin-prose" followed by IDL get assert_equals: expected "" but got "atin-prose"
PASS input.inputMode: setAttribute() to "LATIN-PROSE" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "LATIN-PROSE" followed by IDL get assert_equals: expected "latin-prose" but got "LATIN-PROSE"
PASS input.inputMode: 3 tests
FAIL input.inputMode: setAttribute() to "xfull-width-latin" followed by IDL get assert_equals: expected "" but got "xfull-width-latin"
PASS input.inputMode: setAttribute() to "full-width-latin\0" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "full-width-latin\0" followed by IDL get assert_equals: expected "" but got "full-width-latin\0"
PASS input.inputMode: setAttribute() to "ull-width-latin" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "ull-width-latin" followed by IDL get assert_equals: expected "" but got "ull-width-latin"
PASS input.inputMode: setAttribute() to "FULL-WIDTH-LATIN" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "FULL-WIDTH-LATIN" followed by IDL get assert_equals: expected "full-width-latin" but got "FULL-WIDTH-LATIN"
PASS input.inputMode: 3 tests
FAIL input.inputMode: setAttribute() to "xkana" followed by IDL get assert_equals: expected "" but got "xkana"
PASS input.inputMode: setAttribute() to "kana\0" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "kana\0" followed by IDL get assert_equals: expected "" but got "kana\0"
PASS input.inputMode: setAttribute() to "ana" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "ana" followed by IDL get assert_equals: expected "" but got "ana"
PASS input.inputMode: setAttribute() to "KANA" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "KANA" followed by IDL get assert_equals: expected "kana" but got "KANA"
PASS input.inputMode: 3 tests
FAIL input.inputMode: setAttribute() to "xkatakana" followed by IDL get assert_equals: expected "" but got "xkatakana"
PASS input.inputMode: setAttribute() to "katakana\0" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "katakana\0" followed by IDL get assert_equals: expected "" but got "katakana\0"
PASS input.inputMode: setAttribute() to "atakana" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "atakana" followed by IDL get assert_equals: expected "" but got "atakana"
PASS input.inputMode: setAttribute() to "KATAKANA" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "KATAKANA" followed by IDL get assert_equals: expected "katakana" but got "KATAKANA"
PASS input.inputMode: 3 tests
FAIL input.inputMode: setAttribute() to "xnumeric" followed by IDL get assert_equals: expected "" but got "xnumeric"
PASS input.inputMode: setAttribute() to "numeric\0" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "numeric\0" followed by IDL get assert_equals: expected "" but got "numeric\0"
PASS input.inputMode: setAttribute() to "umeric" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "umeric" followed by IDL get assert_equals: expected "" but got "umeric"
PASS input.inputMode: setAttribute() to "NUMERIC" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "NUMERIC" followed by IDL get assert_equals: expected "numeric" but got "NUMERIC"
PASS input.inputMode: 3 tests
FAIL input.inputMode: setAttribute() to "xtel" followed by IDL get assert_equals: expected "" but got "xtel"
PASS input.inputMode: setAttribute() to "tel\0" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "tel\0" followed by IDL get assert_equals: expected "" but got "tel\0"
PASS input.inputMode: setAttribute() to "el" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "el" followed by IDL get assert_equals: expected "" but got "el"
PASS input.inputMode: setAttribute() to "TEL" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "TEL" followed by IDL get assert_equals: expected "tel" but got "TEL"
PASS input.inputMode: 3 tests
FAIL input.inputMode: setAttribute() to "xemail" followed by IDL get assert_equals: expected "" but got "xemail"
PASS input.inputMode: setAttribute() to "email\0" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "email\0" followed by IDL get assert_equals: expected "" but got "email\0"
PASS input.inputMode: setAttribute() to "mail" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "mail" followed by IDL get assert_equals: expected "" but got "mail"
PASS input.inputMode: setAttribute() to "EMAIL" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "EMAIL" followed by IDL get assert_equals: expected "email" but got "EMAIL"
PASS input.inputMode: 3 tests
FAIL input.inputMode: setAttribute() to "xurl" followed by IDL get assert_equals: expected "" but got "xurl"
PASS input.inputMode: setAttribute() to "url\0" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "url\0" followed by IDL get assert_equals: expected "" but got "url\0"
PASS input.inputMode: setAttribute() to "rl" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "rl" followed by IDL get assert_equals: expected "" but got "rl"
PASS input.inputMode: setAttribute() to "URL" followed by getAttribute() 
FAIL input.inputMode: setAttribute() to "URL" followed by IDL get assert_equals: expected "url" but got "URL"
PASS input.inputMode: 5 tests
FAIL input.inputMode: IDL set to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " followed by IDL get assert_equals: expected "" but got " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo "
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to undefined followed by IDL get assert_equals: expected "" but got "undefined"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to 7 followed by IDL get assert_equals: expected "" but got "7"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to 1.5 followed by IDL get assert_equals: expected "" but got "1.5"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to true followed by IDL get assert_equals: expected "" but got "true"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to false followed by IDL get assert_equals: expected "" but got "false"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to object "[object Object]" followed by IDL get assert_equals: expected "" but got "[object Object]"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to NaN followed by IDL get assert_equals: expected "" but got "NaN"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to Infinity followed by IDL get assert_equals: expected "" but got "Infinity"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to -Infinity followed by IDL get assert_equals: expected "" but got "-Infinity"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "\0" followed by IDL get assert_equals: expected "" but got "\0"
PASS input.inputMode: IDL set to null should not throw 
FAIL input.inputMode: IDL set to null followed by IDL get assert_equals: expected "" but got "null"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to object "test-toString" followed by IDL get assert_equals: expected "" but got "test-toString"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to object "test-valueOf" followed by IDL get assert_equals: expected "" but got "test-valueOf"
PASS input.inputMode: 5 tests
FAIL input.inputMode: IDL set to "xverbatim" followed by IDL get assert_equals: expected "" but got "xverbatim"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "verbatim\0" followed by IDL get assert_equals: expected "" but got "verbatim\0"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "erbatim" followed by IDL get assert_equals: expected "" but got "erbatim"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "VERBATIM" followed by IDL get assert_equals: expected "verbatim" but got "VERBATIM"
PASS input.inputMode: 5 tests
FAIL input.inputMode: IDL set to "xlatin" followed by IDL get assert_equals: expected "" but got "xlatin"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "latin\0" followed by IDL get assert_equals: expected "" but got "latin\0"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "atin" followed by IDL get assert_equals: expected "" but got "atin"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "LATIN" followed by IDL get assert_equals: expected "latin" but got "LATIN"
PASS input.inputMode: 5 tests
FAIL input.inputMode: IDL set to "xlatin-name" followed by IDL get assert_equals: expected "" but got "xlatin-name"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "latin-name\0" followed by IDL get assert_equals: expected "" but got "latin-name\0"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "atin-name" followed by IDL get assert_equals: expected "" but got "atin-name"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "LATIN-NAME" followed by IDL get assert_equals: expected "latin-name" but got "LATIN-NAME"
PASS input.inputMode: 5 tests
FAIL input.inputMode: IDL set to "xlatin-prose" followed by IDL get assert_equals: expected "" but got "xlatin-prose"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "latin-prose\0" followed by IDL get assert_equals: expected "" but got "latin-prose\0"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "atin-prose" followed by IDL get assert_equals: expected "" but got "atin-prose"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "LATIN-PROSE" followed by IDL get assert_equals: expected "latin-prose" but got "LATIN-PROSE"
PASS input.inputMode: 5 tests
FAIL input.inputMode: IDL set to "xfull-width-latin" followed by IDL get assert_equals: expected "" but got "xfull-width-latin"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "full-width-latin\0" followed by IDL get assert_equals: expected "" but got "full-width-latin\0"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "ull-width-latin" followed by IDL get assert_equals: expected "" but got "ull-width-latin"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "FULL-WIDTH-LATIN" followed by IDL get assert_equals: expected "full-width-latin" but got "FULL-WIDTH-LATIN"
PASS input.inputMode: 5 tests
FAIL input.inputMode: IDL set to "xkana" followed by IDL get assert_equals: expected "" but got "xkana"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "kana\0" followed by IDL get assert_equals: expected "" but got "kana\0"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "ana" followed by IDL get assert_equals: expected "" but got "ana"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "KANA" followed by IDL get assert_equals: expected "kana" but got "KANA"
PASS input.inputMode: 5 tests
FAIL input.inputMode: IDL set to "xkatakana" followed by IDL get assert_equals: expected "" but got "xkatakana"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "katakana\0" followed by IDL get assert_equals: expected "" but got "katakana\0"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "atakana" followed by IDL get assert_equals: expected "" but got "atakana"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "KATAKANA" followed by IDL get assert_equals: expected "katakana" but got "KATAKANA"
PASS input.inputMode: 5 tests
FAIL input.inputMode: IDL set to "xnumeric" followed by IDL get assert_equals: expected "" but got "xnumeric"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "numeric\0" followed by IDL get assert_equals: expected "" but got "numeric\0"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "umeric" followed by IDL get assert_equals: expected "" but got "umeric"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "NUMERIC" followed by IDL get assert_equals: expected "numeric" but got "NUMERIC"
PASS input.inputMode: 5 tests
FAIL input.inputMode: IDL set to "xtel" followed by IDL get assert_equals: expected "" but got "xtel"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "tel\0" followed by IDL get assert_equals: expected "" but got "tel\0"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "el" followed by IDL get assert_equals: expected "" but got "el"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "TEL" followed by IDL get assert_equals: expected "tel" but got "TEL"
PASS input.inputMode: 5 tests
FAIL input.inputMode: IDL set to "xemail" followed by IDL get assert_equals: expected "" but got "xemail"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "email\0" followed by IDL get assert_equals: expected "" but got "email\0"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "mail" followed by IDL get assert_equals: expected "" but got "mail"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "EMAIL" followed by IDL get assert_equals: expected "email" but got "EMAIL"
PASS input.inputMode: 5 tests
FAIL input.inputMode: IDL set to "xurl" followed by IDL get assert_equals: expected "" but got "xurl"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "url\0" followed by IDL get assert_equals: expected "" but got "url\0"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "rl" followed by IDL get assert_equals: expected "" but got "rl"
PASS input.inputMode: 2 tests
FAIL input.inputMode: IDL set to "URL" followed by IDL get assert_equals: expected "url" but got "URL"
PASS input.max: 77 tests
PASS input.maxLength: 17 tests
FAIL input.maxLength: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 524288
PASS input.maxLength: 101 tests
FAIL input.maxLength: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 524288
PASS input.min: 77 tests
PASS input.multiple: 79 tests
PASS input.name: 77 tests
PASS input.pattern: 77 tests
PASS input.placeholder: 77 tests
PASS input.readOnly: 79 tests
PASS input.required: 79 tests
PASS input.size: 114 tests
FAIL input.size: IDL set to 2147483648 followed by getAttribute() assert_equals: expected "20" but got "0"
PASS input.size: 2 tests
FAIL input.size: IDL set to 4294967295 followed by getAttribute() assert_equals: expected "20" but got "0"
PASS input.size: IDL set to 4294967295 followed by IDL get 
PASS input.src: 92 tests
PASS input.step: 77 tests
PASS input.type: 103 tests
FAIL input.type: setAttribute() to "datetime" followed by IDL get assert_equals: expected "datetime" but got "text"
PASS input.type: 7 tests
FAIL input.type: setAttribute() to "DATETIME" followed by IDL get assert_equals: expected "datetime" but got "text"
PASS input.type: 301 tests
FAIL input.type: IDL set to "datetime" followed by IDL get assert_equals: expected "datetime" but got "text"
PASS input.type: 11 tests
FAIL input.type: IDL set to "DATETIME" followed by IDL get assert_equals: expected "datetime" but got "text"
PASS input.type: 225 tests
PASS input.defaultValue (<input value>): 77 tests
PASS input.align: 77 tests
PASS input.useMap: 77 tests
PASS button.title: 77 tests
PASS button.lang: 77 tests
PASS button.dir: 151 tests
PASS button.className (<button class>): 77 tests
PASS button.hidden: 79 tests
PASS button.accessKey: 77 tests
PASS button.tabIndex: 10 tests
FAIL button.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS button.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL button.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS button.tabIndex: 36 tests
FAIL button.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS button.tabIndex: 2 tests
FAIL button.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS button.autofocus: 79 tests
PASS button.disabled: 79 tests
PASS button.formEnctype: 151 tests
PASS button.formMethod: 151 tests
PASS button.formNoValidate: 79 tests
PASS button.formTarget: 77 tests
PASS button.name: 77 tests
PASS button.type: 151 tests
PASS button.value: 77 tests
PASS select.title: 77 tests
PASS select.lang: 77 tests
PASS select.dir: 151 tests
PASS select.className (<select class>): 77 tests
PASS select.hidden: 79 tests
PASS select.accessKey: 77 tests
PASS select.tabIndex: 10 tests
FAIL select.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS select.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL select.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS select.tabIndex: 36 tests
FAIL select.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS select.tabIndex: 2 tests
FAIL select.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS select.autofocus: 79 tests
PASS select.disabled: 79 tests
PASS select.multiple: 79 tests
PASS select.name: 77 tests
PASS select.required: 79 tests
PASS select.size: 2 tests
FAIL select.size: setAttribute() to -2147483649 followed by getAttribute() assert_equals: expected "-2147483649" but got "0"
PASS select.size: setAttribute() to -2147483649 followed by IDL get 
FAIL select.size: setAttribute() to -2147483648 followed by getAttribute() assert_equals: expected "-2147483648" but got "0"
PASS select.size: setAttribute() to -2147483648 followed by IDL get 
FAIL select.size: setAttribute() to -36 followed by getAttribute() assert_equals: expected "-36" but got "0"
PASS select.size: setAttribute() to -36 followed by IDL get 
FAIL select.size: setAttribute() to -1 followed by getAttribute() assert_equals: expected "-1" but got "0"
PASS select.size: 10 tests
FAIL select.size: setAttribute() to 2147483648 followed by IDL get assert_equals: expected 0 but got 2147483648
PASS select.size: setAttribute() to 4294967295 followed by getAttribute() 
FAIL select.size: setAttribute() to 4294967295 followed by IDL get assert_equals: expected 0 but got 4294967295
FAIL select.size: setAttribute() to 4294967296 followed by getAttribute() assert_equals: expected "4294967296" but got "0"
PASS select.size: setAttribute() to 4294967296 followed by IDL get 
FAIL select.size: setAttribute() to "" followed by getAttribute() assert_equals: expected "" but got "0"
PASS select.size: setAttribute() to "" followed by IDL get 
FAIL select.size: setAttribute() to "-1" followed by getAttribute() assert_equals: expected "-1" but got "0"
PASS select.size: setAttribute() to "-1" followed by IDL get 
FAIL select.size: setAttribute() to "-0" followed by getAttribute() assert_equals: expected "-0" but got "0"
PASS select.size: 5 tests
FAIL select.size: setAttribute() to "\t7" followed by getAttribute() assert_equals: expected "\t7" but got "7"
PASS select.size: setAttribute() to "\t7" followed by IDL get 
FAIL select.size: setAttribute() to "\v7" followed by getAttribute() assert_equals: expected "\v7" but got "7"
FAIL select.size: setAttribute() to "\v7" followed by IDL get assert_equals: expected 0 but got 7
FAIL select.size: setAttribute() to "\f7" followed by getAttribute() assert_equals: expected "\f7" but got "7"
PASS select.size: setAttribute() to "\f7" followed by IDL get 
FAIL select.size: setAttribute() to " 7" followed by getAttribute() assert_equals: expected " 7" but got "7"
PASS select.size: setAttribute() to " 7" followed by IDL get 
FAIL select.size: setAttribute() to " 7" followed by getAttribute() assert_equals: expected " 7" but got "0"
PASS select.size: setAttribute() to " 7" followed by IDL get 
FAIL select.size: setAttribute() to "﻿7" followed by getAttribute() assert_equals: expected "﻿7" but got "0"
PASS select.size: setAttribute() to "﻿7" followed by IDL get 
FAIL select.size: setAttribute() to "\n7" followed by getAttribute() assert_equals: expected "\n7" but got "7"
PASS select.size: setAttribute() to "\n7" followed by IDL get 
FAIL select.size: setAttribute() to "\r7" followed by getAttribute() assert_equals: expected "\r7" but got "7"
PASS select.size: setAttribute() to "\r7" followed by IDL get 
FAIL select.size: setAttribute() to " 7" followed by getAttribute() assert_equals: expected " 7" but got "7"
FAIL select.size: setAttribute() to " 7" followed by IDL get assert_equals: expected 0 but got 7
FAIL select.size: setAttribute() to " 7" followed by getAttribute() assert_equals: expected " 7" but got "0"
PASS select.size: setAttribute() to " 7" followed by IDL get 
FAIL select.size: setAttribute() to " 7" followed by getAttribute() assert_equals: expected " 7" but got "7"
FAIL select.size: setAttribute() to " 7" followed by IDL get assert_equals: expected 0 but got 7
FAIL select.size: setAttribute() to "᠎7" followed by getAttribute() assert_equals: expected "᠎7" but got "0"
PASS select.size: setAttribute() to "᠎7" followed by IDL get 
FAIL select.size: setAttribute() to " 7" followed by getAttribute() assert_equals: expected " 7" but got "7"
FAIL select.size: setAttribute() to " 7" followed by IDL get assert_equals: expected 0 but got 7
FAIL select.size: setAttribute() to " 7" followed by getAttribute() assert_equals: expected " 7" but got "7"
FAIL select.size: setAttribute() to " 7" followed by IDL get assert_equals: expected 0 but got 7
FAIL select.size: setAttribute() to " 7" followed by getAttribute() assert_equals: expected " 7" but got "7"
FAIL select.size: setAttribute() to " 7" followed by IDL get assert_equals: expected 0 but got 7
FAIL select.size: setAttribute() to " 7" followed by getAttribute() assert_equals: expected " 7" but got "7"
FAIL select.size: setAttribute() to " 7" followed by IDL get assert_equals: expected 0 but got 7
FAIL select.size: setAttribute() to " 7" followed by getAttribute() assert_equals: expected " 7" but got "7"
FAIL select.size: setAttribute() to " 7" followed by IDL get assert_equals: expected 0 but got 7
FAIL select.size: setAttribute() to " 7" followed by getAttribute() assert_equals: expected " 7" but got "7"
FAIL select.size: setAttribute() to " 7" followed by IDL get assert_equals: expected 0 but got 7
FAIL select.size: setAttribute() to " 7" followed by getAttribute() assert_equals: expected " 7" but got "7"
FAIL select.size: setAttribute() to " 7" followed by IDL get assert_equals: expected 0 but got 7
FAIL select.size: setAttribute() to " 7" followed by getAttribute() assert_equals: expected " 7" but got "7"
FAIL select.size: setAttribute() to " 7" followed by IDL get assert_equals: expected 0 but got 7
FAIL select.size: setAttribute() to " 7" followed by getAttribute() assert_equals: expected " 7" but got "7"
FAIL select.size: setAttribute() to " 7" followed by IDL get assert_equals: expected 0 but got 7
FAIL select.size: setAttribute() to " 7" followed by getAttribute() assert_equals: expected " 7" but got "7"
FAIL select.size: setAttribute() to " 7" followed by IDL get assert_equals: expected 0 but got 7
FAIL select.size: setAttribute() to " 7" followed by getAttribute() assert_equals: expected " 7" but got "7"
FAIL select.size: setAttribute() to " 7" followed by IDL get assert_equals: expected 0 but got 7
FAIL select.size: setAttribute() to " 7" followed by getAttribute() assert_equals: expected " 7" but got "0"
PASS select.size: setAttribute() to " 7" followed by IDL get 
FAIL select.size: setAttribute() to "　7" followed by getAttribute() assert_equals: expected "　7" but got "7"
FAIL select.size: setAttribute() to "　7" followed by IDL get assert_equals: expected 0 but got 7
FAIL select.size: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " followed by getAttribute() assert_equals: expected " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " but got "0"
PASS select.size: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " followed by IDL get 
FAIL select.size: setAttribute() to undefined followed by getAttribute() assert_equals: expected "undefined" but got "0"
PASS select.size: setAttribute() to undefined followed by IDL get 
FAIL select.size: setAttribute() to 1.5 followed by getAttribute() assert_equals: expected "1.5" but got "1"
PASS select.size: setAttribute() to 1.5 followed by IDL get 
FAIL select.size: setAttribute() to true followed by getAttribute() assert_equals: expected "true" but got "0"
PASS select.size: setAttribute() to true followed by IDL get 
FAIL select.size: setAttribute() to false followed by getAttribute() assert_equals: expected "false" but got "0"
PASS select.size: setAttribute() to false followed by IDL get 
FAIL select.size: setAttribute() to object "[object Object]" followed by getAttribute() assert_equals: expected "[object Object]" but got "0"
PASS select.size: setAttribute() to object "[object Object]" followed by IDL get 
FAIL select.size: setAttribute() to NaN followed by getAttribute() assert_equals: expected "NaN" but got "0"
PASS select.size: setAttribute() to NaN followed by IDL get 
FAIL select.size: setAttribute() to Infinity followed by getAttribute() assert_equals: expected "Infinity" but got "0"
PASS select.size: setAttribute() to Infinity followed by IDL get 
FAIL select.size: setAttribute() to -Infinity followed by getAttribute() assert_equals: expected "-Infinity" but got "0"
PASS select.size: setAttribute() to -Infinity followed by IDL get 
FAIL select.size: setAttribute() to "\0" followed by getAttribute() assert_equals: expected "\0" but got "0"
PASS select.size: 3 tests
FAIL select.size: setAttribute() to object "3" followed by getAttribute() assert_equals: expected "[object Object]" but got "0"
PASS select.size: 22 tests
PASS datalist.title: 77 tests
PASS datalist.lang: 77 tests
PASS datalist.dir: 151 tests
PASS datalist.className (<datalist class>): 77 tests
PASS datalist.hidden: 79 tests
PASS datalist.accessKey: 77 tests
PASS datalist.tabIndex: 10 tests
FAIL datalist.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS datalist.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL datalist.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS datalist.tabIndex: 36 tests
FAIL datalist.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS datalist.tabIndex: 2 tests
FAIL datalist.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS optgroup.title: 77 tests
PASS optgroup.lang: 77 tests
PASS optgroup.dir: 151 tests
PASS optgroup.className (<optgroup class>): 77 tests
PASS optgroup.hidden: 79 tests
PASS optgroup.accessKey: 77 tests
PASS optgroup.tabIndex: 10 tests
FAIL optgroup.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS optgroup.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL optgroup.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS optgroup.tabIndex: 36 tests
FAIL optgroup.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS optgroup.tabIndex: 2 tests
FAIL optgroup.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS optgroup.disabled: 79 tests
PASS optgroup.label: 77 tests
PASS option.title: 77 tests
PASS option.lang: 77 tests
PASS option.dir: 151 tests
PASS option.className (<option class>): 77 tests
PASS option.hidden: 79 tests
PASS option.accessKey: 77 tests
PASS option.tabIndex: 10 tests
FAIL option.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS option.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL option.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS option.tabIndex: 36 tests
FAIL option.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS option.tabIndex: 2 tests
FAIL option.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS option.disabled: 79 tests
PASS option.label: 47 tests
PASS option.defaultSelected (<option selected>): 79 tests
PASS option.value: 47 tests
PASS textarea.title: 77 tests
PASS textarea.lang: 77 tests
PASS textarea.dir: 151 tests
PASS textarea.className (<textarea class>): 77 tests
PASS textarea.hidden: 79 tests
PASS textarea.accessKey: 77 tests
PASS textarea.tabIndex: 10 tests
FAIL textarea.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS textarea.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL textarea.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS textarea.tabIndex: 36 tests
FAIL textarea.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS textarea.tabIndex: 2 tests
FAIL textarea.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS textarea.autofocus: 79 tests
PASS textarea.cols: 17 tests
FAIL textarea.cols: setAttribute() to 2147483648 followed by IDL get assert_equals: expected 20 but got 2147483648
PASS textarea.cols: setAttribute() to 4294967295 followed by getAttribute() 
FAIL textarea.cols: setAttribute() to 4294967295 followed by IDL get assert_equals: expected 20 but got 4294967295
PASS textarea.cols: 96 tests
FAIL textarea.cols: IDL set to 2147483648 followed by getAttribute() assert_equals: expected "20" but got "0"
PASS textarea.cols: 2 tests
FAIL textarea.cols: IDL set to 4294967295 followed by getAttribute() assert_equals: expected "20" but got "0"
PASS textarea.cols: IDL set to 4294967295 followed by IDL get 
PASS textarea.dirName: 77 tests
PASS textarea.disabled: 79 tests
PASS textarea.inputMode: 5 tests
FAIL textarea.inputMode: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " followed by IDL get assert_equals: expected "" but got " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo "
PASS textarea.inputMode: setAttribute() to undefined followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to undefined followed by IDL get assert_equals: expected "" but got "undefined"
PASS textarea.inputMode: setAttribute() to 7 followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to 7 followed by IDL get assert_equals: expected "" but got "7"
PASS textarea.inputMode: setAttribute() to 1.5 followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to 1.5 followed by IDL get assert_equals: expected "" but got "1.5"
PASS textarea.inputMode: setAttribute() to true followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to true followed by IDL get assert_equals: expected "" but got "true"
PASS textarea.inputMode: setAttribute() to false followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to false followed by IDL get assert_equals: expected "" but got "false"
PASS textarea.inputMode: setAttribute() to object "[object Object]" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to object "[object Object]" followed by IDL get assert_equals: expected "" but got "[object Object]"
PASS textarea.inputMode: setAttribute() to NaN followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to NaN followed by IDL get assert_equals: expected "" but got "NaN"
PASS textarea.inputMode: setAttribute() to Infinity followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to Infinity followed by IDL get assert_equals: expected "" but got "Infinity"
PASS textarea.inputMode: setAttribute() to -Infinity followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to -Infinity followed by IDL get assert_equals: expected "" but got "-Infinity"
PASS textarea.inputMode: setAttribute() to "\0" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "\0" followed by IDL get assert_equals: expected "" but got "\0"
PASS textarea.inputMode: setAttribute() to null followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to null followed by IDL get assert_equals: expected "" but got "null"
PASS textarea.inputMode: setAttribute() to object "test-toString" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to object "test-toString" followed by IDL get assert_equals: expected "" but got "test-toString"
PASS textarea.inputMode: setAttribute() to object "test-valueOf" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to object "test-valueOf" followed by IDL get assert_equals: expected "" but got "test-valueOf"
PASS textarea.inputMode: 3 tests
FAIL textarea.inputMode: setAttribute() to "xverbatim" followed by IDL get assert_equals: expected "" but got "xverbatim"
PASS textarea.inputMode: setAttribute() to "verbatim\0" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "verbatim\0" followed by IDL get assert_equals: expected "" but got "verbatim\0"
PASS textarea.inputMode: setAttribute() to "erbatim" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "erbatim" followed by IDL get assert_equals: expected "" but got "erbatim"
PASS textarea.inputMode: setAttribute() to "VERBATIM" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "VERBATIM" followed by IDL get assert_equals: expected "verbatim" but got "VERBATIM"
PASS textarea.inputMode: 3 tests
FAIL textarea.inputMode: setAttribute() to "xlatin" followed by IDL get assert_equals: expected "" but got "xlatin"
PASS textarea.inputMode: setAttribute() to "latin\0" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "latin\0" followed by IDL get assert_equals: expected "" but got "latin\0"
PASS textarea.inputMode: setAttribute() to "atin" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "atin" followed by IDL get assert_equals: expected "" but got "atin"
PASS textarea.inputMode: setAttribute() to "LATIN" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "LATIN" followed by IDL get assert_equals: expected "latin" but got "LATIN"
PASS textarea.inputMode: 3 tests
FAIL textarea.inputMode: setAttribute() to "xlatin-name" followed by IDL get assert_equals: expected "" but got "xlatin-name"
PASS textarea.inputMode: setAttribute() to "latin-name\0" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "latin-name\0" followed by IDL get assert_equals: expected "" but got "latin-name\0"
PASS textarea.inputMode: setAttribute() to "atin-name" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "atin-name" followed by IDL get assert_equals: expected "" but got "atin-name"
PASS textarea.inputMode: setAttribute() to "LATIN-NAME" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "LATIN-NAME" followed by IDL get assert_equals: expected "latin-name" but got "LATIN-NAME"
PASS textarea.inputMode: 3 tests
FAIL textarea.inputMode: setAttribute() to "xlatin-prose" followed by IDL get assert_equals: expected "" but got "xlatin-prose"
PASS textarea.inputMode: setAttribute() to "latin-prose\0" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "latin-prose\0" followed by IDL get assert_equals: expected "" but got "latin-prose\0"
PASS textarea.inputMode: setAttribute() to "atin-prose" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "atin-prose" followed by IDL get assert_equals: expected "" but got "atin-prose"
PASS textarea.inputMode: setAttribute() to "LATIN-PROSE" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "LATIN-PROSE" followed by IDL get assert_equals: expected "latin-prose" but got "LATIN-PROSE"
PASS textarea.inputMode: 3 tests
FAIL textarea.inputMode: setAttribute() to "xfull-width-latin" followed by IDL get assert_equals: expected "" but got "xfull-width-latin"
PASS textarea.inputMode: setAttribute() to "full-width-latin\0" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "full-width-latin\0" followed by IDL get assert_equals: expected "" but got "full-width-latin\0"
PASS textarea.inputMode: setAttribute() to "ull-width-latin" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "ull-width-latin" followed by IDL get assert_equals: expected "" but got "ull-width-latin"
PASS textarea.inputMode: setAttribute() to "FULL-WIDTH-LATIN" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "FULL-WIDTH-LATIN" followed by IDL get assert_equals: expected "full-width-latin" but got "FULL-WIDTH-LATIN"
PASS textarea.inputMode: 3 tests
FAIL textarea.inputMode: setAttribute() to "xkana" followed by IDL get assert_equals: expected "" but got "xkana"
PASS textarea.inputMode: setAttribute() to "kana\0" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "kana\0" followed by IDL get assert_equals: expected "" but got "kana\0"
PASS textarea.inputMode: setAttribute() to "ana" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "ana" followed by IDL get assert_equals: expected "" but got "ana"
PASS textarea.inputMode: setAttribute() to "KANA" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "KANA" followed by IDL get assert_equals: expected "kana" but got "KANA"
PASS textarea.inputMode: 3 tests
FAIL textarea.inputMode: setAttribute() to "xkatakana" followed by IDL get assert_equals: expected "" but got "xkatakana"
PASS textarea.inputMode: setAttribute() to "katakana\0" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "katakana\0" followed by IDL get assert_equals: expected "" but got "katakana\0"
PASS textarea.inputMode: setAttribute() to "atakana" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "atakana" followed by IDL get assert_equals: expected "" but got "atakana"
PASS textarea.inputMode: setAttribute() to "KATAKANA" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "KATAKANA" followed by IDL get assert_equals: expected "katakana" but got "KATAKANA"
PASS textarea.inputMode: 3 tests
FAIL textarea.inputMode: setAttribute() to "xnumeric" followed by IDL get assert_equals: expected "" but got "xnumeric"
PASS textarea.inputMode: setAttribute() to "numeric\0" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "numeric\0" followed by IDL get assert_equals: expected "" but got "numeric\0"
PASS textarea.inputMode: setAttribute() to "umeric" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "umeric" followed by IDL get assert_equals: expected "" but got "umeric"
PASS textarea.inputMode: setAttribute() to "NUMERIC" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "NUMERIC" followed by IDL get assert_equals: expected "numeric" but got "NUMERIC"
PASS textarea.inputMode: 3 tests
FAIL textarea.inputMode: setAttribute() to "xtel" followed by IDL get assert_equals: expected "" but got "xtel"
PASS textarea.inputMode: setAttribute() to "tel\0" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "tel\0" followed by IDL get assert_equals: expected "" but got "tel\0"
PASS textarea.inputMode: setAttribute() to "el" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "el" followed by IDL get assert_equals: expected "" but got "el"
PASS textarea.inputMode: setAttribute() to "TEL" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "TEL" followed by IDL get assert_equals: expected "tel" but got "TEL"
PASS textarea.inputMode: 3 tests
FAIL textarea.inputMode: setAttribute() to "xemail" followed by IDL get assert_equals: expected "" but got "xemail"
PASS textarea.inputMode: setAttribute() to "email\0" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "email\0" followed by IDL get assert_equals: expected "" but got "email\0"
PASS textarea.inputMode: setAttribute() to "mail" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "mail" followed by IDL get assert_equals: expected "" but got "mail"
PASS textarea.inputMode: setAttribute() to "EMAIL" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "EMAIL" followed by IDL get assert_equals: expected "email" but got "EMAIL"
PASS textarea.inputMode: 3 tests
FAIL textarea.inputMode: setAttribute() to "xurl" followed by IDL get assert_equals: expected "" but got "xurl"
PASS textarea.inputMode: setAttribute() to "url\0" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "url\0" followed by IDL get assert_equals: expected "" but got "url\0"
PASS textarea.inputMode: setAttribute() to "rl" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "rl" followed by IDL get assert_equals: expected "" but got "rl"
PASS textarea.inputMode: setAttribute() to "URL" followed by getAttribute() 
FAIL textarea.inputMode: setAttribute() to "URL" followed by IDL get assert_equals: expected "url" but got "URL"
PASS textarea.inputMode: 5 tests
FAIL textarea.inputMode: IDL set to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " followed by IDL get assert_equals: expected "" but got " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo "
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to undefined followed by IDL get assert_equals: expected "" but got "undefined"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to 7 followed by IDL get assert_equals: expected "" but got "7"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to 1.5 followed by IDL get assert_equals: expected "" but got "1.5"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to true followed by IDL get assert_equals: expected "" but got "true"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to false followed by IDL get assert_equals: expected "" but got "false"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to object "[object Object]" followed by IDL get assert_equals: expected "" but got "[object Object]"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to NaN followed by IDL get assert_equals: expected "" but got "NaN"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to Infinity followed by IDL get assert_equals: expected "" but got "Infinity"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to -Infinity followed by IDL get assert_equals: expected "" but got "-Infinity"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "\0" followed by IDL get assert_equals: expected "" but got "\0"
PASS textarea.inputMode: IDL set to null should not throw 
FAIL textarea.inputMode: IDL set to null followed by IDL get assert_equals: expected "" but got "null"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to object "test-toString" followed by IDL get assert_equals: expected "" but got "test-toString"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to object "test-valueOf" followed by IDL get assert_equals: expected "" but got "test-valueOf"
PASS textarea.inputMode: 5 tests
FAIL textarea.inputMode: IDL set to "xverbatim" followed by IDL get assert_equals: expected "" but got "xverbatim"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "verbatim\0" followed by IDL get assert_equals: expected "" but got "verbatim\0"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "erbatim" followed by IDL get assert_equals: expected "" but got "erbatim"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "VERBATIM" followed by IDL get assert_equals: expected "verbatim" but got "VERBATIM"
PASS textarea.inputMode: 5 tests
FAIL textarea.inputMode: IDL set to "xlatin" followed by IDL get assert_equals: expected "" but got "xlatin"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "latin\0" followed by IDL get assert_equals: expected "" but got "latin\0"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "atin" followed by IDL get assert_equals: expected "" but got "atin"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "LATIN" followed by IDL get assert_equals: expected "latin" but got "LATIN"
PASS textarea.inputMode: 5 tests
FAIL textarea.inputMode: IDL set to "xlatin-name" followed by IDL get assert_equals: expected "" but got "xlatin-name"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "latin-name\0" followed by IDL get assert_equals: expected "" but got "latin-name\0"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "atin-name" followed by IDL get assert_equals: expected "" but got "atin-name"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "LATIN-NAME" followed by IDL get assert_equals: expected "latin-name" but got "LATIN-NAME"
PASS textarea.inputMode: 5 tests
FAIL textarea.inputMode: IDL set to "xlatin-prose" followed by IDL get assert_equals: expected "" but got "xlatin-prose"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "latin-prose\0" followed by IDL get assert_equals: expected "" but got "latin-prose\0"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "atin-prose" followed by IDL get assert_equals: expected "" but got "atin-prose"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "LATIN-PROSE" followed by IDL get assert_equals: expected "latin-prose" but got "LATIN-PROSE"
PASS textarea.inputMode: 5 tests
FAIL textarea.inputMode: IDL set to "xfull-width-latin" followed by IDL get assert_equals: expected "" but got "xfull-width-latin"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "full-width-latin\0" followed by IDL get assert_equals: expected "" but got "full-width-latin\0"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "ull-width-latin" followed by IDL get assert_equals: expected "" but got "ull-width-latin"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "FULL-WIDTH-LATIN" followed by IDL get assert_equals: expected "full-width-latin" but got "FULL-WIDTH-LATIN"
PASS textarea.inputMode: 5 tests
FAIL textarea.inputMode: IDL set to "xkana" followed by IDL get assert_equals: expected "" but got "xkana"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "kana\0" followed by IDL get assert_equals: expected "" but got "kana\0"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "ana" followed by IDL get assert_equals: expected "" but got "ana"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "KANA" followed by IDL get assert_equals: expected "kana" but got "KANA"
PASS textarea.inputMode: 5 tests
FAIL textarea.inputMode: IDL set to "xkatakana" followed by IDL get assert_equals: expected "" but got "xkatakana"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "katakana\0" followed by IDL get assert_equals: expected "" but got "katakana\0"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "atakana" followed by IDL get assert_equals: expected "" but got "atakana"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "KATAKANA" followed by IDL get assert_equals: expected "katakana" but got "KATAKANA"
PASS textarea.inputMode: 5 tests
FAIL textarea.inputMode: IDL set to "xnumeric" followed by IDL get assert_equals: expected "" but got "xnumeric"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "numeric\0" followed by IDL get assert_equals: expected "" but got "numeric\0"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "umeric" followed by IDL get assert_equals: expected "" but got "umeric"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "NUMERIC" followed by IDL get assert_equals: expected "numeric" but got "NUMERIC"
PASS textarea.inputMode: 5 tests
FAIL textarea.inputMode: IDL set to "xtel" followed by IDL get assert_equals: expected "" but got "xtel"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "tel\0" followed by IDL get assert_equals: expected "" but got "tel\0"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "el" followed by IDL get assert_equals: expected "" but got "el"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "TEL" followed by IDL get assert_equals: expected "tel" but got "TEL"
PASS textarea.inputMode: 5 tests
FAIL textarea.inputMode: IDL set to "xemail" followed by IDL get assert_equals: expected "" but got "xemail"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "email\0" followed by IDL get assert_equals: expected "" but got "email\0"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "mail" followed by IDL get assert_equals: expected "" but got "mail"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "EMAIL" followed by IDL get assert_equals: expected "email" but got "EMAIL"
PASS textarea.inputMode: 5 tests
FAIL textarea.inputMode: IDL set to "xurl" followed by IDL get assert_equals: expected "" but got "xurl"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "url\0" followed by IDL get assert_equals: expected "" but got "url\0"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "rl" followed by IDL get assert_equals: expected "" but got "rl"
PASS textarea.inputMode: 2 tests
FAIL textarea.inputMode: IDL set to "URL" followed by IDL get assert_equals: expected "url" but got "URL"
PASS textarea.maxLength: 120 tests
PASS textarea.name: 77 tests
PASS textarea.placeholder: 77 tests
PASS textarea.readOnly: 79 tests
PASS textarea.required: 79 tests
PASS textarea.rows: 17 tests
FAIL textarea.rows: setAttribute() to 2147483648 followed by IDL get assert_equals: expected 2 but got 2147483648
PASS textarea.rows: setAttribute() to 4294967295 followed by getAttribute() 
FAIL textarea.rows: setAttribute() to 4294967295 followed by IDL get assert_equals: expected 2 but got 4294967295
PASS textarea.rows: 96 tests
FAIL textarea.rows: IDL set to 2147483648 followed by getAttribute() assert_equals: expected "2" but got "0"
PASS textarea.rows: 2 tests
FAIL textarea.rows: IDL set to 4294967295 followed by getAttribute() assert_equals: expected "2" but got "0"
PASS textarea.rows: IDL set to 4294967295 followed by IDL get 
PASS textarea.wrap: 77 tests
PASS keygen.title: 77 tests
PASS keygen.lang: 77 tests
PASS keygen.dir: 151 tests
PASS keygen.className (<keygen class>): 77 tests
PASS keygen.hidden: 79 tests
PASS keygen.accessKey: 77 tests
PASS keygen.tabIndex: 10 tests
FAIL keygen.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS keygen.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL keygen.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS keygen.tabIndex: 36 tests
FAIL keygen.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS keygen.tabIndex: 2 tests
FAIL keygen.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS keygen.autofocus: 79 tests
PASS keygen.challenge: 77 tests
PASS keygen.disabled: 79 tests
PASS keygen.keytype: 46 tests
PASS keygen.name: 77 tests
PASS output.title: 77 tests
PASS output.lang: 77 tests
PASS output.dir: 151 tests
PASS output.className (<output class>): 77 tests
PASS output.hidden: 79 tests
PASS output.accessKey: 77 tests
PASS output.tabIndex: 10 tests
FAIL output.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS output.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL output.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS output.tabIndex: 36 tests
FAIL output.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS output.tabIndex: 2 tests
FAIL output.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS output.name: 77 tests
PASS progress.title: 77 tests
PASS progress.lang: 77 tests
PASS progress.dir: 151 tests
PASS progress.className (<progress class>): 77 tests
PASS progress.hidden: 79 tests
PASS progress.accessKey: 77 tests
PASS progress.tabIndex: 10 tests
FAIL progress.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS progress.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL progress.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS progress.tabIndex: 36 tests
FAIL progress.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS progress.tabIndex: 2 tests
FAIL progress.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS meter.title: 77 tests
PASS meter.lang: 77 tests
PASS meter.dir: 151 tests
PASS meter.className (<meter class>): 77 tests
PASS meter.hidden: 79 tests
PASS meter.accessKey: 77 tests
PASS meter.tabIndex: 10 tests
FAIL meter.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS meter.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL meter.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS meter.tabIndex: 36 tests
FAIL meter.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS meter.tabIndex: 2 tests
FAIL meter.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
Harness: the test ran to completion.
